+++ src/main/java/com/sellect/server/order/controller/v4/OrderControllerV4.java
@@ -24,7 +24,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderControllerV4 {
 
     private final OrderServiceV4 orderService;
-    private final UserRepository userRepository; // for test
 
     @PostMapping("/order/payment/{orderId}/ready")
     public ApiResponse<Void> readyPayment(
@@ -33,17 +32,6 @@ public class OrderControllerV4 {
         @RequestParam(name = "coupon_id", required = false) Long userReceivedCouponId) {
         orderService.prepareOrder(user.getId(), orderId, userReceivedCouponId);
         return ApiResponse.ok();
-    }
-
-    @GetMapping("/order/payment/{orderId}/redirect-url")
-    public ApiResponse<PaymentUrlRetrieveResponse> retrievePaymentUrl(
-        @AuthUser User user,
-        @PathVariable Long orderId) {
-        String paymentUrl = orderService.getPaymentUrl(user, orderId);
-        return ApiResponse.ok(PaymentUrlRetrieveResponse.builder()
-            .paymentUrl(paymentUrl)
-            .urlRetrieved(paymentUrl == null ? Boolean.FALSE : Boolean.TRUE)
-            .build());
     }
 
     // === test 용 === //
@@ -55,19 +43,5 @@ public class OrderControllerV4 {
         @RequestParam(name = "coupon_id", required = false) Long userReceivedCouponId) {
         orderService.prepareOrder(userId, orderId, userReceivedCouponId);
         return ApiResponse.ok();
-    }
-
-    @GetMapping("/test/order/payment/{orderId}/redirect-url/{userId}")
-    public ApiResponse<PaymentUrlRetrieveResponse> retrievePaymentUrl(
-        @PathVariable Long userId,
-        @PathVariable Long orderId) {
-        // for test
-        User user = userRepository.findById(userId)
-            .orElseThrow(() -> new CommonException(BError.NOT_EXIST, "user"));
-        String paymentUrl = orderService.getPaymentUrl(user, orderId);
-        return ApiResponse.ok(PaymentUrlRetrieveResponse.builder()
-            .paymentUrl(paymentUrl)
-            .urlRetrieved(paymentUrl == null ? Boolean.FALSE : Boolean.TRUE)
-            .build());
     }
 }
